refacto(frontend): Trying to refacto components used for TodoView
