+++ frontend/src/App.vue
@@ -7,7 +7,7 @@
     <TheAppBar />
 
     <v-main>
-      <router-view></router-view>
+      <router-view />
     </v-main>
   </v-app>
 </template>
@@ -16,11 +16,6 @@
   header {
     line-height: 1.5;
     max-height: 100vh;
-  }
-
-  .logo {
-    display: block;
-    margin: 0 auto 2rem;
   }
 
   nav {
