+++ frontend/src/assets/main.css
@@ -2,7 +2,6 @@
 
 #app {
   max-width: 1280px;
-  margin: 0 auto;
   padding: 2rem;
   font-weight: normal;
 }
@@ -25,12 +24,5 @@ a,
   body {
     display: flex;
     place-items: center;
-  }
-
-  #app {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding: 0 2rem;
   }
 }
