+++ frontend/src/composables/useLeftSidebar.ts
@@ -1,0 +1,14 @@
+import { ref } from 'vue';
+
+const isLeftSidebarOpen = ref(false);
+
+export function useLeftSidebar() {
+  const toggleSidebar = (): void => {
+    isLeftSidebarOpen.value = !isLeftSidebarOpen.value;
+  };
+
+  return {
+    isLeftSidebarOpen,
+    toggleSidebar,
+  };
+}
